+++ domain/src/main/java/com/sitexa/android/domain/User.java
@@ -26,7 +26,7 @@ public class User {
 
     //显示数据拼音的首字母
     private String sortletters;
-    private long userId;
+    private int userId;
     private String mobileNo;
     private String userName;
     private String email;
@@ -58,7 +58,7 @@ public class User {
     private String IMEI;
     private String Password;
 
-    public User(long userId) {
+    public User(int userId) {
         this.userId = userId;
     }
 
@@ -70,11 +70,11 @@ public class User {
         this.sortletters = sortletters;
     }
 
-    public long getUserId() {
+    public int getUserId() {
         return userId;
     }
 
-    public void setUserId(long userId) {
+    public void setUserId(int userId) {
         this.userId = userId;
     }
